fixed checkstyle warnings
