+++ src/main/java/epam/training/hw4/components/Benefit.java
@@ -1,4 +1,4 @@
-package epam.training.hw5.pageComponents;
+package epam.training.hw4.components;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,11 +8,26 @@ import org.openqa.selenium.support.FindBy;
 
 public class Benefit extends AbstractComponent {
 
+    @FindBy(css = "span[class^=\"icons-benefit\"]")
+    private List<WebElement> benefitIcons;
+
     @FindBy(css = "span[class=\"benefit-txt\"]")
     private List<WebElement> benefitTexts;
 
     public Benefit(WebDriver webDriver) {
         super(webDriver);
+    }
+
+    public int getBenefitIconsNumber() {
+        return benefitIcons.size();
+    }
+
+    public int getBenefitTextsNumber() {
+        return benefitTexts.size();
+    }
+
+    public List<WebElement> getBenefitIcons() {
+        return benefitIcons;
     }
 
     public List<String> getBenefitTextsExtracted() {
@@ -21,5 +36,4 @@ public class Benefit extends AbstractComponent {
                 .map(WebElement::getText)
                 .collect(Collectors.toList());
     }
-
 }
