+++ src/main/java/epam/training/hw4/components/NavigationBar.java
@@ -1,4 +1,4 @@
-package epam.training.hw5.pageComponents;
+package epam.training.hw4.components;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -6,9 +6,9 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
+
 @FindBy(css = "ul.sidebar-menu")
 public class NavigationBar extends AbstractComponent {
-
     @FindBy(css = "ul.sidebar-menu>li")
     private List<WebElement> navigationBar;
 
@@ -16,10 +16,18 @@ public class NavigationBar extends AbstractComponent {
         super(webDriver);
     }
 
+    public List<WebElement> getNavigationBar() {
+        return navigationBar;
+    }
+
     public List<String> getNavigationBarExtracted() {
         return navigationBar
                 .stream()
                 .map(WebElement::getText)
                 .collect(Collectors.toList());
+    }
+
+    public int getNavigationBarNumber() {
+        return navigationBar.size();
     }
 }
