+++ src/main/java/epam/training/hw4/pages/DifferentElementsPage.java
@@ -1,7 +1,7 @@
 package epam.training.hw4.pages;
 
-import epam.training.hw4.pageComponents.DifferentElements;
-import epam.training.hw4.pageComponents.LogPanel;
+import epam.training.hw4.components.DifferentElements;
+import epam.training.hw4.components.LogPanel;
 import org.openqa.selenium.WebDriver;
 
 public class DifferentElementsPage extends AbstractPage {
