+++ src/main/java/epam/training/hw5/components/Benefit.java
@@ -1,4 +1,4 @@
-package epam.training.hw4.pageComponents;
+package epam.training.hw5.components;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,26 +8,11 @@ import org.openqa.selenium.support.FindBy;
 
 public class Benefit extends AbstractComponent {
 
-    @FindBy(css = "span[class^=\"icons-benefit\"]")
-    private List<WebElement> benefitIcons;
-
     @FindBy(css = "span[class=\"benefit-txt\"]")
     private List<WebElement> benefitTexts;
 
     public Benefit(WebDriver webDriver) {
         super(webDriver);
-    }
-
-    public int getBenefitIconsNumber() {
-        return benefitIcons.size();
-    }
-
-    public int getBenefitTextsNumber() {
-        return benefitTexts.size();
-    }
-
-    public List<WebElement> getBenefitIcons() {
-        return benefitIcons;
     }
 
     public List<String> getBenefitTextsExtracted() {
@@ -36,4 +21,5 @@ public class Benefit extends AbstractComponent {
                 .map(WebElement::getText)
                 .collect(Collectors.toList());
     }
+
 }
