+++ src/main/java/epam/training/hw5/components/LogPanel.java
@@ -1,10 +1,10 @@
-package epam.training.hw4.pageComponents;
+package epam.training.hw5.components;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+
 
 @FindBy(css = "ul.logs")
 public class LogPanel extends AbstractComponent {
@@ -15,11 +15,12 @@ public class LogPanel extends AbstractComponent {
         super(webDriver);
     }
 
-    public List<String> findLogs(String logText) {
+    public String findLog(String logText) {
         return logList
                 .stream()
                 .map(WebElement::getText)
                 .filter(log -> log.contains(logText))
-                .collect(Collectors.toList());
+                .findFirst()
+                .toString();
     }
 }
